chore: some controller
change middleware logic (remove role middleware due bugged)
