+++ app/Http/Controllers/Api/Audit/SystemController.php
@@ -14,7 +14,6 @@ class SystemController extends ApiController
      */
     public function __construct()
     {
-        $this->middleware(['role:superadmin']);
         $this->middleware(['permission:system.index'], ['only' => ['index']]);
     }
