+++ app/Http/Controllers/Api/Main/TransactionController.php
@@ -4,8 +4,6 @@ namespace App\Http\Controllers\Api\Main;
 
 use App\Http\Controllers\ApiController;
 use App\Services\Api\Main\TransactionService;
-use App\Http\Requests\Api\Main\Transaction\StoreRequest;
-use App\Http\Requests\Api\Main\Transaction\UpdateRequest;
 
 class TransactionController extends ApiController
 {
@@ -16,13 +14,8 @@ class TransactionController extends ApiController
      */
     public function __construct()
     {
-        $this->middleware(['role:admin']);
-        $this->middleware(['role:superadmin']);
         $this->middleware(['permission:transaction.index'], ['only' => ['index']]);
-        $this->middleware(['permission:transaction.store'], ['only' => ['store']]);
         $this->middleware(['permission:transaction.show'], ['only' => ['show']]);
-        $this->middleware(['permission:transaction.update'], ['only' => ['update']]);
-        $this->middleware(['permission:transaction.delete'], ['only' => ['delete']]);
     }
 
     /**
@@ -40,21 +33,6 @@ class TransactionController extends ApiController
     }
 
     /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
-    public function store(StoreRequest $request, TransactionService $service)
-    {
-        try {
-            return $service->store($request->validated());
-        } catch (\Throwable $th) {
-            return $this->catchError($th);
-        }
-    }
-
-    /**
      * Display the specified resource.
      *
      * @param  int  $id
@@ -64,37 +42,6 @@ class TransactionController extends ApiController
     {
         try {
             return $service->show($id);
-        } catch (\Throwable $th) {
-            return $this->catchError($th);
-        }
-    }
-
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function update(UpdateRequest $request, TransactionService $service, $id)
-    {
-        try {
-            return $service->update($request->validated(), $id);
-        } catch (\Throwable $th) {
-            return $this->catchError($th);
-        }
-    }
-
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function destroy(TransactionService $service, $id)
-    {
-        try {
-            return $service->destroy($id);
         } catch (\Throwable $th) {
             return $this->catchError($th);
         }
